Prefactor: Move `get_hash` from docker backend to `strutil` (#18715)
Prefactor for https://github.com/pantsbuild/pants/pull/18708
