+++ src/python/pants/notes/2.16.x.md
@@ -2,6 +2,7 @@
 
 ## What's New
 
+
 ### BUILD files
 
 The new `env` function in BUILD files allows access to environment variables in BUILD files.
@@ -13,6 +14,11 @@ The visibility rules syntax now supports matching on target names. The rules glo
 the same syntax as target selectors.
 
 ### Backends
+
+#### Docker
+
+The [`{pants.hash}`](https://www.pantsbuild.org/docs/tagging-docker-images#string-interpolation-using-placeholder-values)
+generation code was changed such that the generated hash for the same input will be different in pants 2.16.
 
 #### Python
