+++ src/python/pants/util/strutil.py
@@ -3,12 +3,19 @@
 
 from __future__ import annotations
 
+import dataclasses
+import hashlib
+import json
 import re
 import shlex
 import textwrap
-from typing import Callable, Iterable, TypeVar
+from collections import abc
+from typing import Any, Callable, Iterable, TypeVar
 
 from typing_extensions import ParamSpec
+
+from pants.engine.internals.native_engine import Digest
+from pants.util.ordered_set import FrozenOrderedSet, OrderedSet
 
 
 def ensure_binary(text_or_binary: bytes | str) -> bytes:
@@ -335,3 +342,39 @@ def docstring(doc: str | Callable[[], str]) -> Callable[[Callable[P, R]], Callab
         return func
 
     return wrapper
+
+
+class _JsonEncoder(json.JSONEncoder):
+    """Allow us to serialize everything, with a fallback on `str()` in case of any esoteric
+    types."""
+
+    def default(self, o):
+        """Return a serializable object for o."""
+        if isinstance(o, abc.Mapping):
+            return dict(o)
+        if isinstance(o, (abc.Sequence, OrderedSet, FrozenOrderedSet)):
+            return list(o)
+
+        # NB: A quick way to embed the type in the hash so that two objects with the same data but
+        # different types produce different hashes.
+        classname = o.__class__.__name__
+        if dataclasses.is_dataclass(o):
+            return {"__class__.__name__": classname, **dataclasses.asdict(o)}
+        if isinstance(o, (Digest,)):
+            return {"__class__.__name__": classname, "fingerprint": o.fingerprint}
+        return super().default(o)
+
+
+def stable_hash(value: Any, *, name: str = "sha256") -> hashlib._Hash:
+    """Attempts to return a stable hash of the value stable across processes.
+
+    "Stable" here means that if `value` is equivalent in multiple invocations (across multiple
+    processes), it should produce the same hash. To that end, what values are accepted are limited
+    in scope.
+    """
+    return hashlib.new(
+        name,
+        json.dumps(
+            value, indent=None, separators=(",", ":"), sort_keys=True, cls=_JsonEncoder
+        ).encode("utf-8"),
+    )
